FIX(symfonyinsights): Project should use return types declaration

+++ src/Controllers/CategoryController.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Controllers;
 
 use App\Lib\Hydrator;
@@ -10,7 +12,7 @@ use App\Repository\CategoryRepository;
 
 class CategoryController extends DefaultController
 {
-    public function displayAdminCategories()
+    public function displayAdminCategories(): void
     {
         $this->auth->checkAdmin();
 
@@ -19,7 +21,7 @@ class CategoryController extends DefaultController
 
         $categoryRepository = new CategoryRepository();
         $totalItems = $categoryRepository->count();
-        
+
         $pagination = new Pagination($totalItems, $itemsPerPage, $currentPage);
 
         $categories = $categoryRepository->getPaginated($itemsPerPage, $pagination->getOffset());
@@ -30,16 +32,15 @@ class CategoryController extends DefaultController
             'categories' => $categories,
             'pagination' => $paginationHtml
         ]);
-
     }
 
-    public function createForm()
+    public function createForm(): void
     {
         $this->auth->checkAdmin();
         echo $this->twig->getTwig()->render('backend/forms/addCategory.twig');
     }
 
-    public function create(array $params)
+    public function create(array $params): void
     {
         $this->auth->checkAdmin();
         if (!isset($params['post']['name'], $params['post']['description'], $params['post']['slug'])) {
@@ -59,7 +60,7 @@ class CategoryController extends DefaultController
         }
     }
 
-    public function updateForm(array $id)
+    public function updateForm(array $id): void
     {
         $this->auth->checkAdmin();
         $categoryId = (int)$id['id'];
@@ -68,7 +69,7 @@ class CategoryController extends DefaultController
         $existingCategory = $categoryRepository->getById($categoryId);
 
         if (!$existingCategory) {
-            header( $_SERVER["SERVER_PROTOCOL"] . '404 Not Found');
+            header($_SERVER["SERVER_PROTOCOL"] . '404 Not Found');
             echo 'La catégorie n\'existe pas 404 not found baby';
         }
 
@@ -78,7 +79,7 @@ class CategoryController extends DefaultController
     }
 
 
-    public function update(array $id)
+    public function update(array $id): void
     {
         $this->auth->checkAdmin();
         $categoryId = (int)$id['id'];
@@ -91,11 +92,11 @@ class CategoryController extends DefaultController
 
             $categoryRepository = new CategoryRepository();
             $category = $categoryRepository->getById($categoryId);
-    
+
             if ($category) {
                 $category = Hydrator::hydrate($postData, $category);
                 $success = $categoryRepository->update($category);
-    
+
                 if (!$success) {
                     throw new \Exception('Impossible de mettre à jour la Catégorie!');
                 } else {
@@ -105,30 +106,28 @@ class CategoryController extends DefaultController
                 throw new \Exception('Catégorie non trouvé.');
             }
         }
-        
     }
 
 
-    public function delete(array $id)
+    public function delete(array $id): void
     {
         $this->auth->checkAdmin();
-        $id = (int)$id['id']; 
+        $id = (int)$id['id'];
 
         $categoryRepository = new CategoryRepository();
 
         $category = $categoryRepository->getById($id);
 
-        if($category->getId() === $id) {
+        if ($category !== null && $category->getId() === $id) {
             $success = $categoryRepository->delete($category);
-        } else {
-            return false;
-        }
 
-        if (!$success) {
-            throw new \Exception('Impossible de supprimer la categorie!');
-        } else {
+            if (!$success) {
+                throw new \Exception('Impossible de supprimer la categorie!');
+            }
+
             HTTPResponse::redirect('/admin/categories');
+        } else {
+            throw new \Exception('Categorie non trouvée ou ID incorrect.');
         }
     }
-
 }
