+++ src/Controllers/PostController.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Controllers;
 
 use App\Entity\Post;
@@ -8,10 +10,11 @@ use App\Lib\Pagination;
 use App\Lib\HTTPResponse;
 use App\Repository\PostRepository;
 use App\Repository\CommentRepository;
+use Exception;
 
 class PostController extends DefaultController
 {
-    public function index()
+    public function index(): void
     {
         $postRepository = new PostRepository();
         $posts = $postRepository->getLast();
@@ -21,7 +24,7 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function displayPosts()
+    public function displayPosts(): void
     {
         $postStatus = "published";
         $itemsPerPage = 9;
@@ -29,7 +32,7 @@ class PostController extends DefaultController
 
         $postRepository = new PostRepository();
         $totalItems = $postRepository->count();
-        
+
         $pagination = new Pagination($totalItems, $itemsPerPage, $currentPage);
 
         $posts = $postRepository->getPaginated($postStatus, $itemsPerPage, $pagination->getOffset());
@@ -42,9 +45,9 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function displayPost(array $params)
+    public function displayPost(array $params): void
     {
-        $postId = $params['id'];
+        $postId = (int) $params['id'];
 
         $postRepository = new PostRepository();
         $post = $postRepository->getById($postId);
@@ -61,7 +64,7 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function displayAdminPosts()
+    public function displayAdminPosts(): void
     {
         $this->auth->checkAdmin();
 
@@ -80,7 +83,7 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function displayAdminValidatedPosts()
+    public function displayAdminValidatedPosts(): void
     {
         $this->auth->checkAdmin();
 
@@ -90,7 +93,7 @@ class PostController extends DefaultController
 
         $postRepository = new PostRepository();
         $totalItems = $postRepository->count();
-        
+
         $pagination = new Pagination($totalItems, $itemsPerPage, $currentPage);
 
         $posts = $postRepository->getPaginated($postStatus, $itemsPerPage, $pagination->getOffset());
@@ -103,7 +106,7 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function displayAdminDraftedPosts()
+    public function displayAdminDraftedPosts(): void
     {
         $this->auth->checkAdmin();
 
@@ -113,7 +116,7 @@ class PostController extends DefaultController
 
         $postRepository = new PostRepository();
         $totalItems = $postRepository->count();
-        
+
         $pagination = new Pagination($totalItems, $itemsPerPage, $currentPage);
 
         $posts = $postRepository->getPaginated($postStatus, $itemsPerPage, $pagination->getOffset());
@@ -126,13 +129,13 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function createForm()
+    public function createForm(): void
     {
         $this->auth->checkAdmin();
         echo $this->twig->getTwig()->render('backend/forms/addPost.twig');
     }
 
-    public function create(array $postData)
+    public function create(array $postData): void
     {
         $this->auth->checkAdmin();
         $postData = $_POST;
@@ -177,7 +180,7 @@ class PostController extends DefaultController
         }
     }
 
-    public function updateForm(array $id)
+    public function updateForm(array $id): void
     {
         $this->auth->checkAdmin();
         $postId = (int)$id['id'];
@@ -195,7 +198,7 @@ class PostController extends DefaultController
         ]);
     }
 
-    public function update(array $id)
+    public function update(array $id): void
     {
         $this->auth->checkAdmin();
         $postId = (int)$id['id'];
@@ -247,7 +250,7 @@ class PostController extends DefaultController
         }
     }
 
-    public function delete(array $id)
+    public function delete(array $id): void
     {
         $this->auth->checkAdmin();
         $id = (int)$id['id'];
@@ -257,20 +260,20 @@ class PostController extends DefaultController
 
         $post = $postRepository->getById($id);
 
-        if ($post->getId() === $id) {
+        if ($post !== null && $post->getId() === $id) {
             $comments = $commentRepository->getAllBy($id);
             foreach ($comments as $comment) {
                 $commentRepository->delete($comment);
             }
             $success = $postRepository->delete($post);
-        } else {
-            return false;
-        }
 
-        if (!$success) {
-            throw new \Exception('Impossible de supprimer l\'article !');
-        } else {
+            if (!$success) {
+                throw new \Exception('Impossible de supprimer l\'article !');
+            }
+
             HTTPResponse::redirect('/admin/posts');
+        } else {
+            throw new \Exception('Le post n\'a pas été trouvé ou l\'ID est incorrect');
         }
     }
 }
